update 0.8 - final

+++ utils/db_api/models/balance.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, ForeignKey, sql
+from sqlalchemy import Integer, Column, ForeignKey, sql, BigInteger
 
 from utils.db_api.base import Base
 
@@ -7,7 +7,7 @@ class Balance(Base):
     __tablename__ = 'balance'
 
     id = Column(Integer(), primary_key=True)
-    telegram_id = Column(Integer(), ForeignKey('users.telegram_id'), unique=True)
+    telegram_id = Column(BigInteger(), ForeignKey('users.telegram_id'), unique=True)
     amount = Column(Integer(), default=0)
 
     query: sql.Select
